requirements.txt

einops==0.3.0
h5py==2.10.0
ml-collections==0.1.1
pandas==1.3.4
path==15.0.0
pathlib2==2.3.5
pathtools==0.1.2
scikit-learn==1.0
scipy==1.5.2
seaborn==0.11.0
sqlalchemy==1.3.20
statsmodels==0.12.0
style==1.1.0
subprocess32==3.5.4
sympy==1.6.2
tables==3.6.1
tabulate==0.8.9
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.10.1
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.1
torch-metrics==1.1.7
tornado==6.2
tqdm==4.50.2
